Fixes per Kazuho's code review.

+++ lib/pembase64.c
@@ -242,12 +242,12 @@ int ptls_base64_decode(const char *text, ptls_base64_decode_state_t *state, ptls
 static int ptls_compare_separator_line(const char *line, const char *begin_or_end, const char *label)
 {
     int ret = strncmp(line, "-----", 5);
-    int text_index = 5;
+    size_t text_index = 5;
 
     if (ret == 0) {
         size_t begin_or_end_length = strlen(begin_or_end);
         ret = strncmp(line + text_index, begin_or_end, begin_or_end_length);
-        text_index += (int) begin_or_end_length;
+        text_index += begin_or_end_length;
     }
 
     if (ret == 0) {
@@ -258,7 +258,7 @@ static int ptls_compare_separator_line(const char *line, const char *begin_or_en
     if (ret == 0) {
         size_t label_length = strlen(label);
         ret = strncmp(line + text_index, label, label_length);
-        text_index += (int) label_length;
+        text_index += label_length;
     }
 
     if (ret == 0) {
